backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ASR
faster-whisper==1.0.1

# Vector store and embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM
google-generativeai==0.8.0
openai==1.12.0  # optional, for OpenAI support

# Task queue
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utils
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
